Move CiftiParcellate from utils to interfaces.
